Moved report to email function inside Report.kt, removed Report from MainActivity,
